Changed 'middleware, query, repository' chain
