+++ ArcadiaParties.API/CustomMiddlewares/DatabaseRolesMiddleware.cs
@@ -27,7 +27,7 @@ namespace ArcadiaParties.API.CustomMiddlewares
 
             newIdentity.AddClaim(new Claim(ClaimTypes.Name, user.Identity.Name));
 
-            var query = new GetUserRolesQuery(user.Identity.Name);
+            var query = new GetUserRolesQuery(context.User);
             var userRoles = await mediator.Send(query);
 
             foreach (var item in userRoles)
