+++ ArcadiaParties.Data.Abstractions/Repositories/IUserRepository.cs
@@ -8,5 +8,6 @@ namespace ArcadiaParties.Data.Abstractions.Repositories
     {
         Task<UserDTO> GetUser(string identity);
         Task<IEnumerable<UserDTO>> GetUsers();
+        Task<IEnumerable<string>> GetUserRoles(string identity);
     }
 }
